Modify tests to reflects changes in model,move test to another directory
